test(common): use uvu
