+++ package.json
@@ -28,11 +28,18 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/svelte": "^3.0.3",
     "@testing-library/user-event": "^13.1.3",
+    "@types/chai": "^4.3.0",
+    "@types/chai-dom": "^0.0.12",
     "@types/jest": "^26.0.20",
+    "@types/jsdom-global": "^3.0.2",
+    "@types/sinon": "^10.0.10",
     "@typescript-eslint/eslint-plugin": "^5.9.1",
     "@typescript-eslint/parser": "^5.9.1",
     "all-contributors-cli": "^6.20.0",
     "babel-jest": "^26.6.3",
+    "c8": "^7.11.0",
+    "chai": "^4.3.6",
+    "chai-dom": "^1.11.0",
     "cross-env": "^7.0.3",
     "eslint": "^7.18.0",
     "eslint-config-prettier": "^8.2.0",
@@ -45,6 +52,8 @@
     "front-matter": "^4.0.2",
     "husky": "^6.0.0",
     "jest": "^26.6.3",
+    "jsdom": "^19.0.0",
+    "jsdom-global": "^3.0.2",
     "mdsvex": "^0.9.0",
     "pinst": "^2.1.4",
     "pnpm": "^6.29.1",
@@ -54,6 +63,7 @@
     "rollup-plugin-bundle-size": "^1.0.3",
     "rollup-plugin-rename-node-modules": "^1.3.1",
     "rollup-plugin-ts": "^2.0.5",
+    "sinon": "^13.0.1",
     "superstruct": "^0.15.0",
     "svelte": "^3.31.0",
     "svelte-focus-on": "^0.1.4",
@@ -63,8 +73,10 @@
     "tippy.js": "^6.0.0",
     "ts-jest": "^26.5.0",
     "tslib": "^2.3.1",
+    "tsm": "^2.2.1",
     "typedoc": "^0.22.10",
     "typescript": "^4.5.5",
+    "uvu": "^0.5.3",
     "vest": "^4.0.1",
     "yup": "^0.32.9",
     "zod": "^1.11.13"
